+++ app-icon.py
@@ -22,7 +22,7 @@ def resize_image(input_file, output_prefix, sizes):
 
 # Define input file and sizes
 input_png = "static/images/icons/app-icon.png"
-sizes = [48, 57, 60, 76, 96, 120, 144, 152, 192, 256, 384, 512]  # List of sizes in pixels
+sizes = [48, 57, 60, 72, 76, 96, 114, 120, 144, 152, 180, 192, 256, 384, 512]  # List of sizes in pixels
 
 # Call function
 resize_image(input_png, "app-icon", sizes)